hw_5/requirements.txt
numpy
pandas
matplotlib.pyplot
seaborn
math
matplotlib.ticker
